Apply stability tweak to Andrew's recent MTRR patch
Date: Sat, 20 Oct 2012 12:52:43 +0100
From: Andrew Paverd <andrew.paverd@cs.ox.ac.uk>
To: Jonathan McCune <jonmccune@cmu.edu>
CC: flickertcb-devel@lists.sourceforge.net,
Subject: Re: [flickertcb-devel] Restoring MTRRs on Intel Platform

Hi all,

Thanks for the feedback and apologies for the crashes caused by the
patch. Unfortunately I haven't been able to replicate that behaviour on
my main test system (Dell Optiplex 990).

However, I've recently managed to try it on an HP dc7800 and it
consistently triggered a reboot as described. I suspect the problem
might be in the restore_mtrrs function (in mtrrs.c) where the wrmsrl
macro is interacting with the counter value of the for loop. The small
patch below seems to fix the problem on my HP dc7800. Does this work for
anyone else?

Andrew